integrate pemission checker into Client
This creates an IdentityClient abstraction which is the start of enabling the
changes which will eventually allow unprivileged users to
be able to get groups information from third party identity
managers. It encapsulates information about how to create
an IdM third party caveat and to retrieve the identity
from declared information in the discharge macaroon.

In a future PR we'll remove the PermChecker and GroupCache types
as they're now redundant.

As a proof of concept, we changed the charmstore code to use
this (PR to come after this lands).